Add link to background blog post to README.
The blog post that led me to this repository is very useful information for understanding the purpose of this image and why you'd use it. This change makes it easy to find for people that come across this repository another way, or want to keep track of the information there when revisiting this repository later.